Move signals into their own package
