+++ source/@lacqueristas/signals/clickAnchor/index.js
@@ -1,9 +1,10 @@
+import {resolveP} from "ramda-extra"
+
 import updateLocation from "../updateLocation"
 
 export default function clickAnchor (href) {
   return function thunk (dispatch) {
-    return Promise
-      .resolve(href)
+    return resolveP(href)
       .then(() => dispatch(updateLocation(href)))
   }
 }
