+++ source/@lacqueristas/signals/mergeResource/index.js
