+++ source/@lacqueristas/signals/storeSelf/index.js
