+++ source/@lacqueristas/signals/updateLocation/index.js
@@ -1,10 +1,9 @@
 import urlParse from "url-parse"
 import {tapP} from "ramda-extra"
-import history from "../../history"
 import updateNavigation from "../updateNavigation"
 
 export default function updateLocation (href) {
-  return function thunk (dispatch) {
+  return function thunk (dispatch, getState, {history}) {
     return Promise
       .resolve(urlParse(href, true))
       .then(tapP(history.push))
